Added custom function types to replace default add, edit and view functionality

+++ packages/react-renderer-dx/src/datacontainers/grid/entitygrid.tsx
@@ -182,8 +182,22 @@ export const EntityGrid = ({ layout, height }: GridProps) => {
     (row: CrudItem, target: Element) => {
       const actions = [];
 
-      if (t) {
-        if (viewAllowed && viewAllowed(row)) {
+      if (t && viewAllowed && editAllowed && dropAllowed && printAllowed && customFunctionAllowed && customFunctions) {          
+        const defaultViewFunction = customFunctions
+          .map(f =>
+            Object.assign({}, f, { row: row, originalTarget: target })
+          )
+          .find(f => f.type === 'default_view' && customFunctionAllowed(f, row));
+
+        const defaultEditFunction = customFunctions
+          .map(f =>
+            Object.assign({}, f, { row: row, originalTarget: target })
+          )
+          .find(f => f.type === 'default_edit' && customFunctionAllowed(f, row));
+
+        if (defaultViewFunction) {
+          actions.push(defaultViewFunction);
+        } else if (viewAllowed && viewAllowed(row)) {
           actions.push({
             id: 'view',
             text: t('datacontainer.actions.show'),
@@ -194,7 +208,9 @@ export const EntityGrid = ({ layout, height }: GridProps) => {
           });
         }
 
-        if (editAllowed && editAllowed(row)) {
+        if (defaultEditFunction) {
+          actions.push(defaultEditFunction);
+        } else if (editAllowed && editAllowed(row)) {
           actions.push({
             id: 'edit',
             text: t('datacontainer.actions.edit'),
@@ -303,8 +319,10 @@ export const EntityGrid = ({ layout, height }: GridProps) => {
             printExecute(e.itemData.row, e.itemData.originalTarget);
             break;
           default: {
-            switch (e.itemData.type) {
+            switch (e.itemData.type) {              
               case 'edit':
+              case 'default_edit':
+              case 'default_view':
                 if (customEdit) {
                   customEdit(e.itemData, [e.itemData.row]);
                 }
@@ -458,36 +476,53 @@ export const EntityGrid = ({ layout, height }: GridProps) => {
 
   const optionButtonClicked = useCallback(
     (button: OptionButtons, data: CrudItem, target: Element) => {
-      switch (button) {
-        case 'add':
-          break;
-        case 'view':
-          viewExecute(data, target);
-          break;
-        case 'edit':
-          editExecute(data, target);
-          break;
-        case 'delete':
-          deleteExecute(data, target);
-          break;
-        case 'print':
-          printExecute(data, target);
-          break;
-        case 'options':
-          actionRows.current = [data];
-          actionMenuExecute(data, target);
-          break;
-        case 'customoptions':
-          actionRows.current = [data];
-          additionalActionMenuExecute(data, target);
-          break;
+      if (customFunctionAllowed && customFunctions && customEdit) {          
+        const defaultViewFunction = customFunctions          
+          .find(f => f.type === 'default_view' && customFunctionAllowed(f, data));
+
+        const defaultEditFunction = customFunctions          
+          .find(f => f.type === 'default_edit' && customFunctionAllowed(f, data));      
+
+        switch (button) {
+          case 'view':
+            if (defaultViewFunction) {
+              customEdit(defaultViewFunction, [data])
+            } else {
+              viewExecute(data, target);
+            }            
+            break;
+          case 'edit':
+            if (defaultEditFunction) {
+              customEdit(defaultEditFunction, [data])
+            } else {
+              editExecute(data, target);
+            }
+            break;
+          case 'delete':
+            deleteExecute(data, target);
+            break;
+          case 'print':
+            printExecute(data, target);
+            break;
+          case 'options':
+            actionRows.current = [data];
+            actionMenuExecute(data, target);
+            break;
+          case 'customoptions':
+            actionRows.current = [data];
+            additionalActionMenuExecute(data, target);
+            break;
+        }
       }
     },
     [
+      customFunctionAllowed,
+      customFunctions,
+      customEdit,
       viewExecute,
       editExecute,
       deleteExecute,
-      printExecute,
+      printExecute,      
       actionMenuExecute,
       additionalActionMenuExecute,
     ]
@@ -495,32 +530,45 @@ export const EntityGrid = ({ layout, height }: GridProps) => {
 
   const optionButtonAllowed = useCallback(
     (button: OptionButtons, data?: CrudItem): boolean => {
-      switch (button) {
-        case 'add':
-          return (addAllowed && addAllowed()) ?? false;
-        case 'view':
-          return (data && viewAllowed && viewAllowed(data)) ?? false;
-        case 'edit':
-          return (data && editAllowed && editAllowed(data)) ?? false;
-        case 'delete':
-          return (data && dropAllowed && dropAllowed(data)) ?? false;
-        case 'print':
-          return (data && printAllowed && printAllowed(data)) ?? false;
-        case 'options':
-          return true;
-        case 'customoptions': {
-          if (data && customFunctions && customFunctionAllowed) {
-            const allowedAdditionalFunctions = customFunctions?.filter(f =>
-              customFunctionAllowed(f, data)
-            );
+      if (addAllowed && viewAllowed && editAllowed && dropAllowed && printAllowed && customFunctions && customFunctionAllowed) {
+        const defaultAddFunction = customFunctions          
+          .find(f => f.type === 'default_add' && customFunctionAllowed(f));
 
-            return allowedAdditionalFunctions?.length > 0;
+        const defaultViewFunction = customFunctions          
+          .find(f => !!data && f.type === 'default_view' && customFunctionAllowed(f, data));
+
+        const defaultEditFunction = customFunctions          
+          .find(f => !!data && f.type === 'default_edit' && customFunctionAllowed(f, data));      
+
+        switch (button) {
+          case 'add':
+            return (!!defaultAddFunction || addAllowed());
+          case 'view':
+            return !!data && (!!defaultViewFunction || viewAllowed(data));
+          case 'edit':
+            return !!data && (!!defaultEditFunction || editAllowed(data));
+          case 'delete':
+            return (data && dropAllowed && dropAllowed(data)) ?? false;
+          case 'print':
+            return (data && printAllowed && printAllowed(data)) ?? false;
+          case 'options':
+            return true;
+          case 'customoptions': {
+            if (data && customFunctions && customFunctionAllowed) {
+              const allowedAdditionalFunctions = customFunctions?.filter(f =>
+                customFunctionAllowed(f, data)
+              );
+
+              return allowedAdditionalFunctions?.length > 0;
+            }
+
+            return false;
           }
-
-          return false;
+          default:
+            return false;
         }
-        default:
-          return false;
+      } else { 
+        return false;
       }
     },
     [
@@ -609,7 +657,12 @@ export const EntityGrid = ({ layout, height }: GridProps) => {
     ) {
       const addItems = [];
 
-      if (addAllowed()) {
+      const defaultAddFunction = customFunctions        
+        .find(f => f.type === 'default_add' && customFunctionAllowed(f));
+
+      if (defaultAddFunction) {
+        addItems.push({ id: defaultAddFunction.id, text: defaultAddFunction.text, customFunction: defaultAddFunction })
+      } else if (addAllowed()) {
         addItems.push({
           id: 'none',
           text: t('datacontainer.actions.add', { entity: displayName }),
